Move interceptors to separate package and move MetricRegistryBean back - it requires dependency on mpmetrics api
