+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/bnd.bnd
@@ -25,6 +25,7 @@ Import-Package: \
 Export-Package: \
 	io.astefanutti.metrics.cdi,\
 	com.ibm.ws.microprofile.metrics.cdi,\
+	com.ibm.ws.microprofile.metrics.cdi.interceptors,\
 	com.ibm.ws.microprofile.metrics.cdi.producer,\
 	com.ibm.ws.microprofile.metrics.cdi.helper
