+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/src/com/ibm/ws/microprofile/metrics/cdi/interceptors/MetricsBinding.java
@@ -21,7 +21,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package io.astefanutti.metrics.cdi;
+package com.ibm.ws.microprofile.metrics.cdi.interceptors;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -33,4 +33,5 @@ import javax.interceptor.InterceptorBinding;
 @InterceptorBinding
 @Target({ ElementType.TYPE })
 @Retention(RetentionPolicy.RUNTIME)
-/* package-private */ @interface MetricsBinding {}
+public @interface MetricsBinding {
+}
