+++ dev/com.ibm.ws.microprofile.metrics.cdi.2.0/src/io/astefanutti/metrics/cdi/MetricRegistryBean.java
@@ -21,7 +21,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.ibm.ws.microprofile.metrics.cdi;
+package io.astefanutti.metrics.cdi;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
